requirements.txt
# Code Review Agent - Dependencies
# Python 3.11+

# Web & API
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# HTTP & Async
httpx>=0.26.0
python-multipart>=0.0.6

# LLM (OpenAI-compatible API)
openai>=1.12.0

# Config & Data
pydantic>=2.5.0
pydantic-settings>=2.1.0
pyyaml>=6.0.1

# Storage (optional SQLite for tasks)
aiosqlite>=0.19.0

# CLI
typer>=0.9.0

# Utils
structlog>=24.1.0
python-dotenv>=1.0.0
